requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
guardrails.ai